+++ src/Components/improvment/improvmentPresenter.tsx
@@ -1,11 +1,11 @@
 import { PageExplanation } from "../PageExplanation";
 import { useHistory } from "react-router-dom";
-import ImprovmentItemPresenter from "./improvmentItem/improvmentItemPresenter";
+import ImprovmentItemPresenter from "./ImprovmentItem/ImprovmentItemPresenter";
 import { useState, useEffect } from "react";
-import { improvement, list } from "./imporvmentContainer";
+import { improvement, list } from "./ImprovmentContainer";
 import * as S from "./styled";
 import * as I from "../../Asset/SVG";
-import * as C from "./imporvmentContainer";
+import * as C from "./ImprovmentContainer";
 import { useRecoilValue } from "recoil";
 import { HasAdminToken } from "Atom";
 import { useModal } from "Context/Modal";
