+++ src/Components/index.tsx
@@ -4,8 +4,8 @@ export { default as StartPage } from "./StartPage/StartPresenter";
 export { default as LeaveCommentsPage } from "./LeaveComments/LeaveCommentPresenter";
 export { default as Top10Page } from "./Top30Page/Top30Presenter";
 export { default as PageExplanation } from "./PageExplanation/PageExplanationContainer";
-export { default as ImprovmentPage } from "./improvment/improvmentPresenter";
-export { default as AboutPage } from "./about/aboutPresenter";
+export { default as ImprovmentPage } from "./Improvment/ImprovmentPresenter";
+export { default as AboutPage } from "./About/AboutPresenter";
 export { default as GoodBtn } from "./GoodBtn/GoodBtnPresenter";
 export { default as IssueBox } from "./IssueBox/IssueBoxPresenter";
 export { ChannelTalkHelmet } from  "./Helmet";
