+++ src/Page/index.tsx
@@ -1,5 +1,5 @@
 export { default as LeaveComments } from "./LeaveComments";
 export { default as Start } from "./Start";
 export { default as Top10 } from "./Top10";
-export { default as about } from "./about";
-export { default as improvment } from "./improvment";
+export { default as About } from "./About";
+export { default as Improvment } from "./Improvment";
